putting it all in the password.tsx file and adding success checkmark
